+++ handle_width.c
@@ -6,7 +6,7 @@
 /*   By: bkandemi <bkandemi@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/22 13:39:30 by bkandemi          #+#    #+#             */
-/*   Updated: 2022/03/24 10:01:55 by bkandemi         ###   ########.fr       */
+/*   Updated: 2022/03/25 09:01:21 by bkandemi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ static int	calc_padding(t_flag *flag, char *str, uintmax_t len, int neg)
 int	handle_width(t_flag *flag, char *str, int negative)
 {
 	int			ret;
-	uintmax_t	len;
+	int	len;
 	int			pad_size;
 
 	ret = 0;
