+++ print_int.c
@@ -6,13 +6,13 @@
 /*   By: bkandemi <bkandemi@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/22 13:50:25 by bkandemi          #+#    #+#             */
-/*   Updated: 2022/03/24 10:49:40 by bkandemi         ###   ########.fr       */
+/*   Updated: 2022/03/25 09:04:10 by bkandemi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-static uintmax_t	calc_len(t_flag *flag, char *str)
+static int	calc_len(t_flag *flag, char *str)
 {
 	if (*str == '0' && flag->dot == TRUE && flag->prec == 0)
 		return (0);
@@ -21,7 +21,7 @@ static uintmax_t	calc_len(t_flag *flag, char *str)
 
 int	handle_int(t_flag *flag, char *str, int negative)
 {
-	uintmax_t	len;
+	int	len;
 	int			ret;
 
 	len = calc_len(flag, str);
